Always pull prefecthq docker images in the helm chart
Changes the default setting from IfNotPresent to Always since we are
using the 'latest' tag. This matches K8s behavior if the pull policy
were null. I considered setting this to null by default and depending
on the K8s default handling depending on the tag value, but it sounds
like the 'latest' -> 'Always' default pull policy behavior is
deprecated and I don't want to rely on it.

We will likely set this back to IfNotPresent when we pin the
prefectVersion for helm chart releases. However, some people prefer
Always because then the image cannot accidentally be grabbed by a
deployment without the proper secrets (not relevant for our images
since they are public)