Support null environment & run_config
Flows serialized by `prefect` should always have an `environment` and/or
a `run_config` present, but this isn't enforced by the schema. This adds
a quick fix to support flows with both `run_config` and `environment` as
`None`, to match the behavior of the cloud api (which already has this
fix).