Adds Cloud SQL Auth Proxy Support
When the postgresql DB is hosted on Google Cloud SQL and hosted on a GKE
Cluster, the can be a need to use the Cloud SQL Auth Proxy.  Using the
Auth Proxy is also the recommended approach to connect to Cloud SQL from
GKE (https://cloud.google.com/sql/docs/mysql/connect-kubernetes-engine)

The mechanism for controlling the upgrading of the database has also
been modified.  .Values.postgresql.upgradeDb takes a boolean value and
is used to control whether the graphql node will run the db upgrade
script or not.